+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 - Sale order: fix unremovable sale order line and editable invoiced sale order line.
 - Fix set a prospect as the client of a sale order.
 - Stock Move: select stock move with current invoice canceled in multiple sale and purchase invoicing form
+- Set trading name automatically when creating or merging purchase order
 
 ## [5.0.0-rc2] - 2018-05-09
 ## Improvements